requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
google-genai>=1.0.0
# Heavy local models - comment out for cloud-only mode
# paddlepaddle>=2.5.0
# paddleocr>=2.7.0
streamlit>=1.29.0
requests>=2.31.0
tqdm>=4.66.0
psutil>=5.9.0
